+++ Client/Scripts/Gameplay/Interfaces/ICamera.cs
@@ -10,5 +10,5 @@ public interface ICamera
     /// True zoom is a numeric value representing coefficient of enlargement of zoomed in objects.
     /// When camera zoom is mentioned anywhere in the project, it refers to this value, not the Godot camera's value.
     /// </summary>
-    float TrueZoom { get; }
+    double TrueZoom { get; }
 }
